chore(org/plain/old/retro/shooter/monitor/Screen.java) refactor render method
